[base] Create hdfUtils.h
